pyproject.toml
[tool.poetry]
name = "note-app"
version = "0.1.0"
description = ""
authors = ["DeepBlockDeepak <jordan.medina1729@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
boto3 = "^1.37.28"
python-dotenv = "^1.1.0"
pydantic = "^2.11.2"
mangum = "^0.19.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
coverage = "^7.8.0"
black = "^25.1.0"
flake8 = "^7.2.0"
aws-cdk-lib = "^2.188.0"
constructs = "^10.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
